(core) Extract attempts and naming helpers from hitta_ansikten
Move log_attempt_stats to core/attempts.py and the pure name/filename
helpers (extract_prefix_suffix, is_unrenamed, collect_persons_for_files,
normalize_name, split_fornamn_efternamn, resolve_fornamn_dubletter,
build_new_filename) to core/naming.py. Function bodies are unchanged;
hitta_ansikten imports back the names it uses. Add core.attempts and
core.naming to the PyInstaller hiddenimports.
